requirements.txt
asgiref==3.5.2
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==4.0.6
django-cors-headers==3.13.0
django-phonenumber-field==6.3.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
executing==0.8.3
idna==3.3
inflection==0.5.1
ipython==8.4.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.7.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
openapi-codec==1.3.2
packaging==21.3
parso==0.8.3
pexpect==4.8.0
phonenumbers==8.12.51
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
swagger-spec-validator==2.7.4
traitlets==5.3.0
uritemplate==4.1.1
urllib3==1.26.10
wcwidth==0.2.5
